requirements.txt
flask
bcrypt
google-generativeai
python-dotenv
